refactor(catalog): introduce pluginfilter type and give filter methods better names
