+++ public/app/features/plugins/admin/helpers.ts
@@ -1,6 +1,6 @@
 import { config } from '@grafana/runtime';
 import { gt } from 'semver';
-import { CatalogPlugin, CatalogPluginDetails, LocalPlugin, Plugin, Version } from './types';
+import { CatalogPlugin, CatalogPluginDetails, LocalPlugin, Plugin, Version, PluginFilter } from './types';
 
 export function isGrafanaAdmin(): boolean {
   return config.bootData.user.isGrafanaAdmin;
@@ -125,3 +125,24 @@ export function getCatalogPluginDetails(
 
   return plugin;
 }
+
+export const isInstalled: PluginFilter = (plugin, query) =>
+  query === 'installed' ? plugin.isInstalled : !plugin.isCore;
+
+export const isType: PluginFilter = (plugin, query) => query === 'all' || plugin.type === query;
+
+export const matchesKeyword: PluginFilter = (plugin, query) => {
+  if (!query) {
+    return true;
+  }
+  const fields: String[] = [];
+  if (plugin.name) {
+    fields.push(plugin.name.toLowerCase());
+  }
+
+  if (plugin.orgName) {
+    fields.push(plugin.orgName.toLowerCase());
+  }
+
+  return fields.some((f) => f.includes(query.toLowerCase()));
+};
